Add nested report generation crawl
Also move super long function into utils file
