+++ server/cli/src/lib.rs
@@ -1,3 +1,5 @@
 extern crate diesel;
 mod refresh_dates;
 pub use refresh_dates::*;
+mod report_utils;
+pub use report_utils::*;
